Fixing positionreport (not tested in SmartCARS yet)
The Api has always created a row instead of updating the existing one. The reason for this behavior was that in the firstOrNew function always the complete bid object was given instead of just the ID. Which led to the fact that the existing row was not found. I've also changed that only the user_id is stored in the $report array instead of the whole object. FirstOrNew now checks next to the bid_user also the user_id (will probably be changed again)